% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropLoops.R
\name{DropLoops}
\alias{DropLoops}
\title{A variety of options to drop loops from 'LoopRanges' objects}
\usage{
DropLoops(loop_ranges, type, size)
}
\arguments{
\item{loop_ranges}{A 'LoopRanges' class object}

\item{type}{Type of filtering when determining which loops to drop:
\itemize{
 \item "Anchor" - Drops loops based on given anchor sizes 
 \item "Loop" - Drops loops based on end-to-end loop size
}}
}
\description{
Drops loops based on a type of filter and given options. Can be used to filter the loops in the 'LoopRanges' object before or after determining consensus
}
