% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElementsToRanges.R
\name{ElementsToRanges}
\alias{ElementsToRanges}
\title{Create a list of ranges objects from element data}
\usage{
ElementsToRanges(..., file_format, element_names = NULL)
}
\arguments{
\item{...}{Up to 10 tab delimited element data files in BED4 or BED12 format}

\item{file_format}{The type of input element file - either 'BED4' or 'BED12' format}

\item{element_names}{A character vector of names for the element datasets (optional)}
}
\value{
A 'ElementRanges' class object: list of GRanges element data objects
}
\description{
Uses tab delimited element data in the form of BED4 or BED12 files to create GRanges element objects.
}
