% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoopsToRanges.R
\name{LoopsToRanges}
\alias{LoopsToRanges}
\title{Create a list of ranges objects from looping data}
\usage{
LoopsToRanges(..., loop_names = NULL)
}
\arguments{
\item{...}{Up to 10 tab delimited looping data files in HiCCUPS output format}

\item{loop_names}{A character vector of names for the loop datasets (optional)}
}
\value{
A 'LoopRanges' class object: list of GRanges looping data objects
}
\description{
Uses tab delimited looping data in the form of BED12 files to create GRanges loop objects
}
