% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropLoops.R
\name{DropLoops}
\alias{DropLoops}
\title{A variety of options to drop loops from the 'LoopRanges' object}
\usage{
DropLoops(type, ...)
}
\arguments{
\item{type}{Type of filtering when determining which loops to drop:
\itemize{
 \item "Anchor Size" - Drops loops based on given anchor sizes as the option.
 \item 
}}

\item{...}{Options for filtering based on method type selected:
\itemize{
 \item For method "Anchor Size" - a single integer or numerical vector of anchor sizes. 
 \item For method "Loop Size" 
}}
}
\description{
Drops loops based on a type of filter and given options. Can be used to filter the loops in the 'LoopRanges' object before or after determining consensus.
}
