% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportBED.R
\name{ExportBED}
\alias{ExportBED}
\title{Export a LoopRanges or \emph{italics}ElementRanges object to a BED file}
\usage{
ExportBED(obj, file)
}
\arguments{
\item{obj}{An object of \emph{italics}LoopRanges or \emph{italics}ElementRanges class}

\item{file}{A string indicating the name and save location of BED file}
}
\description{
Uses ElementRanges objects to output to BED format, and LoopRanges objects to output to BEDPE format
}
