% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedElements.R
\name{LinkedElements}
\alias{LinkedElements}
\title{Determines elements linked by loop anchors}
\usage{
LinkedElements(loop_ranges, element_ranges_1, element_ranges_2,
  overlap_threshold = 1)
}
\arguments{
\item{loop_ranges}{A single GRanges loop object in the form of GRangesList class. Should ideally be the consensus loops object created through ConsensusLoops().}

\item{element_ranges_1}{The first element ranges to be considered. If using an 'ElementRanges' class object, subset by using list indexing.}

\item{element_ranges_2}{The second element ranges to be considered. If using an 'ElementRanges' class object, subset by using list indexing.}

\item{overlap_threshold}{Single numerical input for significant base-pair overlap to be considered 'overlapping'. Default is 1 base-pair.}
}
\value{
Returns a dataframe indicating the unique element links present in the form of their first metadata columns. The first element ranges nodes are always under one column, and the second under the other.
}
\description{
Using a finalized loops object and two sets of elements, determines which elements are linked by loops through overlap of anchor regions
}
